Wrapper for vcfR2genind function that assigns pops

Wrapper for vcfR2genind function that assigns pops

Usage

vcf_to_genind(x, pops = NULL)

Arguments

x

can either be an object of class 'vcfR' or a path to a .vcf file

pops

if NULL (default), and there are no pops detected from the vcf, each individual is assigned its own pop. If FALSE then genind$pop is left NULL. Alternatively, a vector of population assignments for each individual can be provided

Value

returns genind object

Examples

data("ex_vcf")
vcf_to_genind(ex_vcf)
#> Warning: no pops were provided, assigning a pop to each individual (to stop this, set pop = FALSE)
#> /// GENIND OBJECT /////////
#> 
#>  // 100 individuals; 1,000 loci; 1,884 alleles; size: 1.3 Mb
#> 
#>  // Basic content
#>    @tab:  100 x 1884 matrix of allele counts
#>    @loc.n.all: number of alleles per locus (range: 1-2)
#>    @loc.fac: locus factor for the 1884 columns of @tab
#>    @all.names: list of allele names for each locus
#>    @ploidy: ploidy of each individual  (range: 2-2)
#>    @type:  codom
#>    @call: adegenet::df2genind(X = t(x), sep = sep)
#> 
#>  // Optional content
#>    @pop: population of each individual (group size range: 1-1)