Sliding window map of genetic diversity

Sliding window map of genetic diversity

Usage

window_gd(
  vcf,
  coords,
  lyr,
  stat = "pi",
  fact = 0,
  wdim = 5,
  rarify = FALSE,
  rarify_n = 4,
  rarify_nit = 5,
  min_n = 2,
  fun = mean,
  parallel = FALSE,
  nloci = NULL
)

Arguments

vcf

object of type vcf ( (note: order matters! the coordinate and genetic data should be in the same order, there are currently no checks for this.))

coords

coordinates (two columns, the first should be x and the second should be y and the order should be the same as the genetic data),

lyr

RasterLayer to slide window across

stat

genetic diversity stat to calculate (can either be "pi" for nucleotide diversity, "het" for average heterozygosity across all loci, "allelic.richness" for average allelic richness across all loci, or "biallelic.richness" to get average allelic richness across all loci for a biallelic dataset (this option faster than "allelic.richness"))

fact

aggregation factor to apply to the RasterLayer (note: increasing this value reduces computational time)

wdim

dimensions (height x width) of window, if only one value is provided a square window is created

rarify

if rarify = TRUE, rarefaction is performed

rarify_n

if rarify = TRUE, number of points to use for rarefaction

rarify_nit

if rarify = TRUE, number of iterations to use for rarefaction

min_n

min number of samples to calculate allelic richness for (equal to rarify_n if provided, otherwise defaults to 2)

fun

function to use to summarize data in window (defaults to base R mean)

parallel

whether to parallelize the function (see vignette for setting up a cluster to do so)

nloci

for calculating pi, if nloci=NULL (default), returns the sum over SNPs of nucleotide diversity; otherwise return the average nucleotide diversity per nucleotide given the length nloci of the sequence (L argument in pi.dosage function)

Value

RasterStack that includes a raster of genetic diversity and a raster of the number of samples within the window for each cell

Examples

load_mini_ex()
#> Error in load_mini_ex(): could not find function "load_mini_ex"
wpi <- window_gd(vcf, coords, lyr, stat = "pi", nloci = 10, rarify_n = 4, rarify_nit = 5, rarify = TRUE)
#> Error in window_gd(vcf, coords, lyr, stat = "pi", nloci = 10, rarify_n = 4,     rarify_nit = 5, rarify = TRUE): object 'vcf' not found
whet <- window_gd(vcf, coords, lyr, stat = "het", nloci = 10, rarify_n = 4, rarify_nit = 5, rarify = TRUE)
#> Error in window_gd(vcf, coords, lyr, stat = "het", nloci = 10, rarify_n = 4,     rarify_nit = 5, rarify = TRUE): object 'vcf' not found
war <- window_gd(vcf, coords, lyr, stat = "biallelic.richness", nloci = 10, rarify_n = 4, rarify_nit = 5, rarify = TRUE)
#> Error in window_gd(vcf, coords, lyr, stat = "biallelic.richness", nloci = 10,     rarify_n = 4, rarify_nit = 5, rarify = TRUE): object 'vcf' not found
plot_gd(wpi)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'wpi' not found
plot_gd(whet)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'whet' not found
plot_gd(war)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'war' not found
plot_count(wpi)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'wpi' not found