Helper function for window_gd

Helper function for window_gd

Usage

window_gd_general(
  gen,
  coords,
  lyr,
  stat = calc_mean_ar,
  fact = 0,
  wdim = 10,
  rarify = FALSE,
  rarify_n = 4,
  rarify_nit = 10,
  min_n = 2,
  fun = mean,
  parallel = FALSE,
  nloci = NULL
)

Arguments

gen

genetic data (note: order matters! the coordinate and genetic data should be in the same order, there are currently no checks for this.)

coords

coordinates (two columns, the first should be x and the second should be y and the order should be the same as the genetic data),

lyr

RasterLayer to slide window across

fact

aggregation factor to apply to the RasterLayer (note: increasing this value reduces computational time)

wdim

dimensions (height x width) of window, if only one value is provided a square window is created

rarify

if rarify = TRUE, rarefaction is performed

rarify_n

if rarify = TRUE, number of points to use for rarefaction

rarify_nit

if rarify = TRUE, number of iterations to use for rarefaction

min_n

min number of samples to calculate allelic richness for (equal to rarify_n if provided, otherwise defaults to 2)

fun

function to use to summarize data in window (defaults to base R mean)

parallel

whether to parallelize the function (see vignette for setting up a cluster to do so)

nloci

for calculating pi, if nloci=NULL (default), returns the sum over SNPs of nucleotide diversity; otherwise return the average nucleotide diversity per nucleotide given the length nloci of the sequence (L argument in pi.dosage function)

Value

RasterStack that includes a raster of genetic diversity and a raster of the number of samples within the window for each cell